+++ kotlinx-coroutines-core/common/test/flow/operators/OnEachTest.kt
@@ -43,7 +43,7 @@ class OnEachTest : TestBase() {
             latch.receive()
             throw TestException()
             it + 1
-        }.onErrorReturn(42)
+        }.catch { emit(42) }
 
         assertEquals(42, flow.single())
         assertTrue(cancelled)